We have switched to Vundle
I know I've been using it for awhile but I updated everything so that it
is actually right. There is a minor change to bashrc that is because I'm
on a mac right now. Ignore that. The major changes are to deal with
vundle and less submodule stuff. All that is a submodule now is
vividchalk.