Stop zsh autorenaming tmux windows
Added a disable to stop zsh from automatically renaming tmux windows.
This will mean that when we do binder-, that the name will stick. SO
suggests adding "set allow-rename off" to ~/.tmux.conf but that results
in an error on the first opening of tmux. Plus, zsh is the issue.
Hopefully this doesn't break other things.