Rebasing
Was going to wait to push this all at once but ehhhh let's be good and
branch this and push our updates as we go.

Rebasing and completely reorganizing my dotfiles. This is becoming a
bigger task but it highly rewarding. Motivation started with having
a better install script after the Nvidia internship. Ansible helped but
we can do better, especially w.r.t. the Mac vs Linux. I'm not going to
bother with Ubuntu 20.04 though as no one should be using it...

Also this helps practice and improve my bash scripting skills!