Updating Vimrc
Made a few things cleaner and a bit better functionality.
Starting to learn some vimscript so this is making a better rcfile.

This includes the leader commands that were file specific now are only
created when we're using that type of file. For example \tex only works
from a tex file and you won't try to pdflatex a python file. We also let
this endup doing \gcc will call gcc or g++ depending if you're in a .c
or .cpp file. And \m runs make in either but the special make if in a
tex!