Installs and Skels
Most for installs is the github installer template.
This is for a record of how to get versions and releases, more of a
guide than template.

For skeletons I added jails for fail2ban.
These can be used as is but should be modified for your environment.
For systemd the timeshift timers can be used, but mount ones should be
edited for your configuration.