Minor notes and Wez Updates
Notes:
Minor to bashing
Linux got notes about problem solving with an nvidia driver issue...
why nvidia why!?
Added some stuff about timeshift and avoiding a second error.
Need to add stuff about systemd boot (should go in systemd) and dracut

Wezterm is now updated to dynamically change the logo for which machine
we're on.  for a mac  for linux.
Need to see if I can get it to run
`lsb_release -i | cut -d ":" -f2 | tr -d '\t'`
And be able to change the icon for whichever distro we're on.
Even better if we can get a marker for an ssh session.