Notes, FZF tmux fix, caffeine
Notes on xargs, some linux, and unfinished tailscale

Install_files/osx.sh is really just a note

zsh/aliases is a fix for the fzf to make sure we show full size chafa
photos when not in a tmux session. We'll check for the $TERM_PROGRAM
variable and check if it is "tmux" to see if we are actively using that
session.

Caffeine script to turn on and off.
Working on integrating into wezterm but this is... not working as
expected?