Wezterm and make readme
Fixed the wezterm command to get it to run an osa script, in this case
to toggle caffeine with opt + c.

The make readme script is mostly complete, but in progress. It is a
template that should be modified to make a readme file or make the table
of contents for a directory wherein you can include this in your readme.