More notes and fuckoff script
The fuckoff script doesn't work as intended.
The issue is with Apple's daemon system.
Cron is not a solution and I'm fighting launchctl with the lack of
debugging and how prolific out of date command syntax is. If only debug
messages were provided or Apple had some decent debugging outputs...